% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mylog.R
\name{my_log}
\alias{my_log}
\title{trustworthy log
returns log(trust me)}
\usage{
my_log(x)
}
\arguments{
\item{x}{numerical vector of length n}
}
\value{
the trustworthy log of x
}
\description{
trustworthy log
returns log(trust me)
}
\examples{
x = 10
mylog(x)
log(10)
}
